Fixed session_id circular import
